+++ SME.CDEP.Aplicacao/Servicos/ServicoImportacaoArquivoAcervoBibliografico.cs
@@ -6,8 +6,8 @@ using SME.CDEP.Aplicacao.DTOS;
 using SME.CDEP.Aplicacao.Servicos.Interface;
 using SME.CDEP.Dominio.Constantes;
 using SME.CDEP.Dominio.Entidades;
-using SME.CDEP.Dominio.Excecoes;
 using SME.CDEP.Dominio.Extensions;
+using SME.CDEP.Infra;
 using SME.CDEP.Infra.Dados.Repositorios.Interfaces;
 using SME.CDEP.Infra.Dominio.Enumerados;
 
@@ -15,19 +15,17 @@ namespace SME.CDEP.Aplicacao.Servicos
 {
     public class ServicoImportacaoArquivoAcervoBibliografico : ServicoImportacaoArquivoBase, IServicoImportacaoArquivoAcervoBibliografico 
     {
-        private readonly IServicoAcervoBibliografico servicoAcervoBibliografico;
-        private readonly IMapper mapper;
-        
+        private readonly IServicoMensageria servicoMensageria;
+
         public ServicoImportacaoArquivoAcervoBibliografico(IRepositorioImportacaoArquivo repositorioImportacaoArquivo, IServicoMaterial servicoMaterial,
-            IServicoEditora servicoEditora,IServicoSerieColecao servicoSerieColecao,IServicoIdioma servicoIdioma, IServicoAssunto servicoAssunto,
-            IServicoCreditoAutor servicoCreditoAutor,IServicoConservacao servicoConservacao, IServicoAcessoDocumento servicoAcessoDocumento,
-            IServicoCromia servicoCromia, IServicoSuporte servicoSuporte,IServicoFormato servicoFormato,IServicoAcervoBibliografico servicoAcervoBibliografico,IMapper mapper,
-            IRepositorioParametroSistema repositorioParametroSistema)
-            : base(repositorioImportacaoArquivo, servicoMaterial, servicoEditora,servicoSerieColecao, servicoIdioma, servicoAssunto, servicoCreditoAutor,
-                servicoConservacao,servicoAcessoDocumento,servicoCromia,servicoSuporte, servicoFormato, mapper,repositorioParametroSistema)
+            IServicoEditora servicoEditora, IServicoSerieColecao servicoSerieColecao, IServicoIdioma servicoIdioma, IServicoAssunto servicoAssunto,
+            IServicoCreditoAutor servicoCreditoAutor, IServicoConservacao servicoConservacao, IServicoAcessoDocumento servicoAcessoDocumento,
+            IServicoCromia servicoCromia, IServicoSuporte servicoSuporte, IServicoFormato servicoFormato, IMapper mapper,
+            IRepositorioParametroSistema repositorioParametroSistema, IServicoMensageria servicoMensageria)
+            : base(repositorioImportacaoArquivo, servicoMaterial, servicoEditora, servicoSerieColecao, servicoIdioma, servicoAssunto, servicoCreditoAutor,
+                servicoConservacao, servicoAcessoDocumento, servicoCromia, servicoSuporte, servicoFormato, mapper, repositorioParametroSistema)
         {
-            this.servicoAcervoBibliografico = servicoAcervoBibliografico ?? throw new ArgumentNullException(nameof(servicoAcervoBibliografico));
-            this.mapper = mapper ?? throw new ArgumentNullException(nameof(mapper));
+            this.servicoMensageria = servicoMensageria ?? throw new ArgumentNullException(nameof(servicoMensageria));
         }
 
         public void DefinirCreditosAutores(List<IdNomeTipoDTO> creditosAutores)
@@ -74,17 +72,6 @@ namespace SME.CDEP.Aplicacao.Servicos
             return await ObterRetornoImportacaoAcervo(arquivoImportado, JsonConvert.DeserializeObject<IEnumerable<AcervoBibliograficoLinhaDTO>>(arquivoImportado.Conteudo), false);
         }
 
-        public async Task CarregarDominiosBibliograficos()
-        {
-            await CarregarTodosOsDominios();
-            
-            await ObterMateriaisPorTipo(TipoMaterial.BIBLIOGRAFICO);
-                
-            await ObterCoAutores(TipoCreditoAutoria.Coautor);
-            
-            await ObterCreditosAutoresPorTipo(TipoCreditoAutoria.Autoria);
-        }
-
         public async Task<ImportacaoArquivoRetornoDTO<AcervoLinhaErroDTO<AcervoBibliograficoDTO,AcervoBibliograficoLinhaRetornoDTO>,AcervoLinhaRetornoSucessoDTO>> ImportarArquivo(IFormFile file)
         {
             ValidarArquivo(file);
@@ -93,20 +80,18 @@ namespace SME.CDEP.Aplicacao.Servicos
 
             var importacaoArquivo = ObterImportacaoArquivoParaSalvar(file.FileName, TipoAcervo.Bibliografico, JsonConvert.SerializeObject(acervosBibliograficosLinhas));
             
-            await CarregarDominiosBibliograficos();
-            
             var importacaoArquivoId = await PersistirImportacao(importacaoArquivo);
+            
+            await servicoMensageria.Publicar(RotasRabbit.ExecutarImportacaoArquivoAcervoBibliograficoUseCase, importacaoArquivoId, Guid.NewGuid());
 
-            ValidarPreenchimentoValorFormatoQtdeCaracteres(acervosBibliograficosLinhas);
-            
-            await AtualizarImportacao(importacaoArquivoId, JsonConvert.SerializeObject(acervosBibliograficosLinhas),ImportacaoStatus.ValidadoPreenchimentoValorFormatoQtdeCaracteres);
-            
-            await PersistenciaAcervo(acervosBibliograficosLinhas);
-            await AtualizarImportacao(importacaoArquivoId, JsonConvert.SerializeObject(acervosBibliograficosLinhas), acervosBibliograficosLinhas.Any(a=> a.PossuiErros) ? ImportacaoStatus.Erros : ImportacaoStatus.Sucesso);
-            
-            var arquivoImportado = await repositorioImportacaoArquivo.ObterPorId(importacaoArquivoId);
-
-            return await ObterRetornoImportacaoAcervo(arquivoImportado, acervosBibliograficosLinhas);
+            return new ImportacaoArquivoRetornoDTO<AcervoLinhaErroDTO<AcervoBibliograficoDTO, AcervoBibliograficoLinhaRetornoDTO>, AcervoLinhaRetornoSucessoDTO>()
+            {
+                Id = importacaoArquivoId,
+                Nome = importacaoArquivo.Nome,
+                TipoAcervo = importacaoArquivo.TipoAcervo,
+                DataImportacao = DateTimeExtension.HorarioBrasilia(),
+                Status = ImportacaoStatus.Pendente
+            };
         }
 
         private async Task<ImportacaoArquivoRetornoDTO<AcervoLinhaErroDTO<AcervoBibliograficoDTO,AcervoBibliograficoLinhaRetornoDTO>,AcervoLinhaRetornoSucessoDTO>> ObterRetornoImportacaoAcervo(ImportacaoArquivo arquivoImportado, IEnumerable<AcervoBibliograficoLinhaDTO> acervosBibliograficosLinhas, bool estaImportandoArquivo = true)
@@ -277,57 +262,6 @@ namespace SME.CDEP.Aplicacao.Servicos
             return mensagemErro.ToArray();
         }
 
-        public async Task PersistenciaAcervo(IEnumerable<AcervoBibliograficoLinhaDTO> acervosBibliograficosLinhas)
-        {
-            foreach (var acervoBibliograficoLinha in acervosBibliograficosLinhas.Where(w=> !w.PossuiErros))
-            {
-                try
-                {
-                    var acervoBibliografico = new AcervoBibliograficoCadastroDTO()
-                    {
-                        Titulo = acervoBibliograficoLinha.Titulo.Conteudo,
-
-                        SubTitulo = acervoBibliograficoLinha.SubTitulo.Conteudo,
-
-                        MaterialId = ObterMaterialBibliograficoIdPorValorDoCampo(acervoBibliograficoLinha.Material.Conteudo),
-
-                        CreditosAutoresIds = ObterCreditoAutoresIdsPorValorDoCampo(acervoBibliograficoLinha.Autor.Conteudo, TipoCreditoAutoria.Autoria),
-
-                        CoAutores = ObterCoAutoresTipoAutoria(acervoBibliograficoLinha.CoAutor.Conteudo,acervoBibliograficoLinha.TipoAutoria.Conteudo),
-
-                        EditoraId = ObterEditoraIdOuNuloPorValorDoCampo(acervoBibliograficoLinha.Editora.Conteudo),
-
-                        AssuntosIds = ObterAssuntosIdsPorValorDoCampo(acervoBibliograficoLinha.Assunto.Conteudo),
-
-                        Ano = acervoBibliograficoLinha.Ano.Conteudo,
-                        Edicao = acervoBibliograficoLinha.Edicao.Conteudo,
-                        NumeroPagina = acervoBibliograficoLinha.NumeroPaginas.Conteudo.ObterInteiroOuNuloPorValorDoCampo(),
-                        Largura = acervoBibliograficoLinha.Largura.Conteudo,
-                        Altura = acervoBibliograficoLinha.Altura.Conteudo,
-
-                        SerieColecaoId = ObterSerieColecaoIdOuNuloPorValorDoCampo(acervoBibliograficoLinha.SerieColecao.Conteudo),
-
-                        Volume = acervoBibliograficoLinha.Volume.Conteudo,
-
-                        IdiomaId = ObterIdiomaIdPorValorDoCampo(acervoBibliograficoLinha.Idioma.Conteudo),
-
-                        LocalizacaoCDD = acervoBibliograficoLinha.LocalizacaoCDD.Conteudo,
-                        LocalizacaoPHA = acervoBibliograficoLinha.LocalizacaoPHA.Conteudo,
-                        NotasGerais = acervoBibliograficoLinha.NotasGerais.Conteudo,
-                        Isbn = acervoBibliograficoLinha.Isbn.Conteudo,
-                        Codigo = acervoBibliograficoLinha.Codigo.Conteudo
-                    };
-                    await servicoAcervoBibliografico.Inserir(acervoBibliografico);
-
-                    acervoBibliograficoLinha.DefinirLinhaComoSucesso();
-                }
-                catch (Exception ex)
-                {
-                    acervoBibliograficoLinha.DefinirLinhaComoErro(ex.Message);
-                }
-            }
-        }
-
         public CoAutorDTO[] ObterCoAutoresTipoAutoria(string coautores, string tiposAutoria)
         {
             if (coautores.NaoEstaPreenchido())
@@ -354,98 +288,6 @@ namespace SME.CDEP.Aplicacao.Servicos
             }).ToArray();
 
             return coAutoresCompletos;
-        }
-
-        public void ValidarPreenchimentoValorFormatoQtdeCaracteres(IEnumerable<AcervoBibliograficoLinhaDTO> linhas)
-        {
-            var autores = CreditosAutores.Select(s=> mapper.Map<IdNomeDTO>(s));
-            var coAutores = CoAutores.Select(s=> mapper.Map<IdNomeDTO>(s));
-            var materiaisBibliograficos = Materiais.Where(w => w.Tipo == (int)TipoMaterial.BIBLIOGRAFICO).Select(s=> mapper.Map<IdNomeDTO>(s));
-            
-            foreach (var linha in linhas)
-            {
-                try
-                {
-                    ValidarPreenchimentoLimiteCaracteres(linha.Titulo, Constantes.TITULO);
-                    ValidarPreenchimentoLimiteCaracteres(linha.SubTitulo, Constantes.SUB_TITULO);
-                    
-                    ValidarPreenchimentoLimiteCaracteres(linha.Material,Constantes.MATERIAL);
-                    ValidarConteudoCampoComDominio(linha.Material, materiaisBibliograficos, Constantes.MATERIAL);
-                    
-                    ValidarPreenchimentoLimiteCaracteres(linha.Autor,Constantes.AUTOR);
-                    ValidarConteudoCampoListaComDominio(linha.Autor, autores, Constantes.AUTOR);
-
-                    ValidarPreenchimentoLimiteCaracteres(linha.CoAutor,Constantes.CO_AUTOR);
-                    ValidarConteudoCampoListaComDominio(linha.CoAutor, coAutores, Constantes.CO_AUTOR);
-
-                    ValidarPreenchimentoLimiteCaracteres(linha.TipoAutoria,Constantes.TIPO_AUTORIA);
-                    
-                    if (linha.TipoAutoria.Conteudo.EstaPreenchido() && linha.CoAutor.Conteudo.NaoEstaPreenchido())
-                        DefinirMensagemErro(linha.TipoAutoria, Constantes.CAMPO_COAUTOR_SEM_PREENCHIMENTO_E_TIPO_AUTORIA_PREENCHIDO);
-
-                    if (linha.TipoAutoria.Conteudo.SplitPipe().Count() > linha.CoAutor.Conteudo.SplitPipe().Count())
-                        DefinirMensagemErro(linha.TipoAutoria, Constantes.TEMOS_MAIS_TIPO_AUTORIA_QUE_COAUTORES);
-                    
-                    ValidarPreenchimentoLimiteCaracteres(linha.Editora,Constantes.EDITORA);
-                    ValidarConteudoCampoComDominio(linha.Editora, Editoras, Constantes.EDITORA);
-                    
-                    ValidarPreenchimentoLimiteCaracteres(linha.Assunto,Constantes.ASSUNTO);
-                    ValidarConteudoCampoListaComDominio(linha.Assunto, Assuntos, Constantes.ASSUNTO);
-
-                    ValidarPreenchimentoLimiteCaracteres(linha.Ano,Constantes.ANO);
-                    ValidarPreenchimentoLimiteCaracteres(linha.Edicao,Constantes.EDICAO);
-                    
-                    ValidarPreenchimentoLimiteCaracteres(linha.NumeroPaginas,Constantes.NUMERO_PAGINAS);
-                    ValidarPreenchimentoLimiteCaracteres(linha.Largura,Constantes.LARGURA);
-                    ValidarPreenchimentoLimiteCaracteres(linha.Altura,Constantes.ALTURA);
-                    
-                    ValidarPreenchimentoLimiteCaracteres(linha.SerieColecao,Constantes.SERIE_COLECAO);
-                    ValidarConteudoCampoListaComDominio(linha.SerieColecao, SeriesColecoes, Constantes.SERIE_COLECAO);
-                    
-                    ValidarPreenchimentoLimiteCaracteres(linha.Volume,Constantes.VOLUME);
-                    
-                    ValidarPreenchimentoLimiteCaracteres(linha.Idioma,Constantes.IDIOMA);
-                    ValidarConteudoCampoComDominio(linha.Idioma, Idiomas, Constantes.IDIOMA);
-                    
-                    ValidarPreenchimentoLimiteCaracteres(linha.LocalizacaoCDD,Constantes.LOCALIZACAO_CDD);
-                    ValidarPreenchimentoLimiteCaracteres(linha.LocalizacaoPHA,Constantes.LOCALIZACAO_PHA);
-                    
-                    ValidarPreenchimentoLimiteCaracteres(linha.NotasGerais,Constantes.NOTAS_GERAIS);
-                    ValidarPreenchimentoLimiteCaracteres(linha.Isbn,Constantes.ISBN);
-                    
-                    ValidarPreenchimentoLimiteCaracteres(linha.Codigo,Constantes.TOMBO);
-                    linha.PossuiErros = PossuiErro(linha);
-                }
-                catch (Exception e)
-                {
-                    linha.DefinirLinhaComoErro(string.Format(Constantes.OCORREU_UMA_FALHA_INESPERADA_NA_LINHA_X_MOTIVO_Y, e.Message));
-                }
-            }
-        }
-
-        private bool PossuiErro(AcervoBibliograficoLinhaDTO linha)
-        {
-            return linha.Titulo.PossuiErro 
-                   || linha.SubTitulo.PossuiErro 
-                   || linha.Material.PossuiErro 
-                   || linha.Autor.PossuiErro 
-                   || linha.CoAutor.PossuiErro 
-                   || linha.TipoAutoria.PossuiErro 
-                   || linha.Editora.PossuiErro 
-                   || linha.Edicao.PossuiErro 
-                   || linha.Assunto.PossuiErro 
-                   || linha.Ano.PossuiErro 
-                   || linha.NumeroPaginas.PossuiErro
-                   || linha.Largura.PossuiErro 
-                   || linha.Altura.PossuiErro 
-                   || linha.SerieColecao.PossuiErro 
-                   || linha.Volume.PossuiErro 
-                   || linha.Idioma.PossuiErro
-                   || linha.LocalizacaoCDD.PossuiErro
-                   || linha.LocalizacaoPHA.PossuiErro
-                   || linha.NotasGerais.PossuiErro
-                   || linha.Isbn.PossuiErro
-                   || linha.Codigo.PossuiErro;
         }
 
         private async Task<IEnumerable<AcervoBibliograficoLinhaDTO>> LerPlanilha(IFormFile file)
