+++ SME.CDEP.Infra.Servicos/Rabbit/RotasRabbit.cs
@@ -19,5 +19,7 @@ namespace SME.CDEP.Infra
         
         public const string NotificacaoDevolucaoEmprestimoAtrasoProlongado = "cdep.emprestimo.situacao.atraso.prolongado.aviso";
         public const string NotificacaoDevolucaoEmprestimoAtrasoProlongadoUsuario = "cdep.emprestimo.situacao.atraso.prolongado.aviso.usuario";
+        
+        public const string ExecutarImportacaoArquivoAcervoBibliograficoUseCase = "cdep.importacao.arquivo.acervo.bibliografico";
     }
 }
